refactor(queries): move organization, team and integrator hooks into the queries layer
